event-log-service/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
SQLAlchemy
psycopg2-binary